Tweak makefile
Tweak the makefile to make it easier to build. Leptonica now defaults
to installing from libs in the standard libdir and includes in the
include dir, rather than from a Leptonica tarball. In addition, the
default EXTRA flags now provide defaults for Homebrew on Mac OS X,
rather than Fink.